OMX: Add new error codes to ignore corrupted header error from DSP
Quant method one is the tool which is available only with ASP profile.
Decoder is detecting this error and populating an error as invalid tool.

There was an optimization done at SN level with respect to buffer handling.
To work the optimization properly it was important to populate the profile
information to the application. We have added these checks for tools to
inform application as there is invalid tool setting, as profile is mentioned
as SP with ASP tools enable. Application can take corrective action
accordingly.

Algorithm is populating M4H3DEC_TI_ERROR_invalidToolsSimpleObject error.
Application can check for this specific error and ignore this error.

This error is populated only to inform application as mismatch in profile
and tool. Application can choose to continue decoding.

DR:OMAPS00243797

Change-Id: Ie0df4cb636059f0b171ca55cd0c5cd5d4a2b8d94
Signed-off-by: Bangoria, Tushar <x0099756@ti.com>
Signed-off-by: Ryan J. Rhee <junhwa.rhee@windriver.com>

Conflicts:

	video/src/openmax_il/video_decode/src/OMX_VideoDec_Utils.c